pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-cracker"
version = "0.1.0"
description = "Proving quantum superposition and its relational imprint on observable metrics"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "qiskit>=1.0",
    "qiskit-aer>=0.13",
    "matplotlib>=3.8",
    "PyOpenGL>=3.1.7",
    "glfw>=2.6",
    "pyrr>=0.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.2",
    "jupyter>=1.0",
    "ipykernel>=6.29",
    "ecdsa>=0.19",
    "pycryptodome>=3.20",
]

[project.scripts]
quantum-cracker = "quantum_cracker.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
